Reactivating yapf and a few more pre-commits
I didn't reactivate prospector yet - this requires
a bit more time and I would like to do it in a different PR,
so this can get merged in the meantime.

I also set the version of this plugin consistently to 2.0.0a1
since I think we want to release a 2.0 out of it soon.