Fixing problems in the documentation
This now works locally.
Also, a test has been added to Travis to ensure that
the documentation builds correctly, and without warnings.
I decided to put this test in the 'pre-commit' set of tests,
because I feel it's a bit too much to have another full set
of tests just for the docs, and I feel in a sense this is a
'pre-commit' test.