Split tests into separate parts for calculation and parser.
Instead of running a single integration test that explicitly calls Wannier90,
the tests are now split into parts for setting up the inputs (the CalcJob)
and parsing the outputs (the Parser). This follows the example from
aiida-quantumespresso. As part of this, the dependency on aiida-pytest is
removed, because the additional features it provides are no longer needed.