Implementing input_filename and output_filename
These are added in the input specs. We dropped the seedname, since
the input_filename and output_filename are "required" by
AiiDA to make `verdi calcjob inputcat` and `verdi calcjob outputcat`
to work. We could keep all three of them (seedname, input_filename
and output_filename), but we would need to validate the consistency
between all three of them,

We therefore infer the seedname from the input filename, and only
validate that it is consistent with the output filename.