Addition of three tests
In the first test we check that the code raises an error if the input filename does not end in .win. The second test check that we get an error if there is an inconsistency between input and output filenames. Finally, the third test check that an other error is raised if we don't find the matching .werr file in the retrieved list.