Remove opening transport in 'prepare_for_submission'
To avoid opening a transport to the 'remote_input_folder', we use
globbing to specify all _possible_ input files that can be given.
Because we can no longer determine if there are overlaps between
local and remote input, the two inputs are now mutually exclusive.
Also, the check for the required input files is performed only
when a local input is passed.
Because passing an input does not make sense in the
'postproc_setup' mode, this is no longer allowed.
The implementation separates the specification of input files
(which are optional / required, should be copied / symlinked),
which is shared between local and remote input, from the
generation of the input lists.

TODO: The remote_input_folder version was tested with local
transport, but should probably also be tested with ssh transport.