Set the `invalidates_cache` keyword for certain exit codes.
For exit codes where no or incomplete output was retrieved, the
`invalidates_cache` keyword is set to True. Since these are
exit codes that are often caused by intermittent issues (out of
memory, node failure, ...), they should not be used as cache
source.
Because the `invalidates_cache` keyword was introduced in AiiDA
version 1.1, support for the 1.0 version is dropped.