Move from Travis CI to Github Actions.
The workflows defined in `.github/workflows` are auto-
generated from `.github/workflows_source`. This has the
advantage that YAML anchors can be used, reducing duplicate
boilerplate.
The `utils/interpolate_yaml_anchors.py` is vendored to do
this generating.

Because the previous version of `prospector` failed with our
updated dependencies, it was updated. This in turn updated
`pylint`, which required some changes to which messages are
ignored.